+++ docker/docker-compose.dev.yml
@@ -50,7 +50,7 @@ services:
     networks:
       - app-network
     healthcheck:
-      test: [ "CMD", "curl", "-f", "http://localhost:5010/health" ]
+      test: [ "CMD", "curl", "-f", "http://localhost:5010/auth/health" ]
       interval: 30s
       timeout: 5s
       retries: 3
@@ -68,7 +68,7 @@ services:
     networks:
       - app-network
     healthcheck:
-      test: [ "CMD", "curl", "-f", "http://localhost:5012/health" ]
+      test: [ "CMD", "curl", "-f", "http://localhost:5012/notification/health" ]
       interval: 30s
       timeout: 5s
       retries: 3
@@ -86,7 +86,7 @@ services:
     networks:
       - app-network
     healthcheck:
-      test: [ "CMD", "curl", "-f", "http://localhost:5013/health" ]
+      test: [ "CMD", "curl", "-f", "http://localhost:5013/docs/health" ]
       interval: 30s
       timeout: 5s
       retries: 3
