+++ nginx/nginx.conf
@@ -1,10 +1,12 @@
 events {}
 
 http {
+    resolver 127.0.0.11 valid=30s;
+
     server {
         listen 80;
-        resolver 127.0.0.11 valid=30s;
-        set $auth_upstream "http://auth:5010/";
+        set $auth_upstream "http://auth:5010";
+        set $notification_upstream "http://notification:5012";
         set $docs_upstream "http://docs:5013";
 
         location /health {
@@ -13,7 +15,7 @@ http {
         }
 
        location /auth/ {
-           proxy_pass         http://auth:5010/;
+           proxy_pass         $auth_upstream;
            proxy_http_version 1.1;
            proxy_set_header   Upgrade $http_upgrade;
            proxy_set_header   Connection keep-alive;
@@ -26,7 +28,7 @@ http {
           allow 172.18.0.0/16;
           deny all;
 
-          proxy_pass         http://notification:5012/;
+          proxy_pass         $notification_upstream;
           proxy_http_version 1.1;
           proxy_set_header   Upgrade $http_upgrade;
           proxy_set_header   Connection keep-alive;
@@ -36,12 +38,17 @@ http {
        }
 
       location /docs/ {
-          proxy_pass http://docs:5013/;
+          proxy_pass $docs_upstream;
           proxy_http_version 1.1;
           proxy_set_header Upgrade $http_upgrade;
           proxy_set_header Connection keep-alive;
           proxy_set_header Host $host;
           proxy_cache_bypass $http_upgrade;
+      }
+
+      location @module_unavailable {
+          default_type application/json;
+          return 503 '{"error": "Service unavailable", "details": "The target module is currently down or unreachable."}';
       }
     }
 }
