+++ notification/src/main.ts
@@ -7,6 +7,7 @@ import { ValidationPipe } from '@nestjs/common';
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   app.useGlobalPipes(new ValidationPipe());
+  app.setGlobalPrefix('notification');
 
   const configService = app.get(ConfigService);
   const appConf: ConfigType<typeof appConfig> = configService.get('app');
